# sql: account for memory of results of window functions computations