@yuzefovich 2019-07-12
Local benchmarks don't show any significant difference (they actually show a slight performance improvement that I would explain by noise):
```
+ make bench PKG=./pkg/sql/distsqlrun BENCHTIMEOUT=5m BENCHES=BenchmarkWindower 'TESTFLAGS=-count 10 -benchmem'
name                                                                          old time/op    new time/op    delta
Windower/SUM()_OVER_()-12                                                       64.5ms ± 0%    63.6ms ± 2%  -1.45%  (p=0.005 n=6+10)
Windower/SUM()_OVER_(ORDER_BY)-12                                                126ms ± 2%     121ms ± 1%  -3.93%  (p=0.000 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/)-12                 172ms ± 2%     169ms ± 1%  -1.91%  (p=0.002 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/_ORDER_BY)-12        190ms ± 2%     183ms ± 1%  -3.39%  (p=0.000 n=10+9)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/)-12             82.0ms ± 2%    81.7ms ± 2%    ~     (p=0.631 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/_ORDER_BY)-12     104ms ± 1%     104ms ± 2%    ~     (p=0.853 n=10+10)

name                                                                          old speed      new speed      delta
Windower/SUM()_OVER_()-12                                                     37.2MB/s ± 0%  37.8MB/s ± 2%  +1.48%  (p=0.005 n=6+10)
Windower/SUM()_OVER_(ORDER_BY)-12                                             19.1MB/s ± 2%  19.8MB/s ± 1%  +4.08%  (p=0.000 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/)-12              14.0MB/s ± 2%  14.2MB/s ± 1%  +1.94%  (p=0.003 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/_ORDER_BY)-12     12.6MB/s ± 2%  13.1MB/s ± 1%  +3.50%  (p=0.000 n=10+9)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/)-12           29.3MB/s ± 2%  29.4MB/s ± 2%    ~     (p=0.617 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/_ORDER_BY)-12  23.1MB/s ± 1%  23.1MB/s ± 3%    ~     (p=0.837 n=10+10)

name                                                                          old alloc/op   new alloc/op   delta
Windower/SUM()_OVER_()-12                                                       25.8MB ± 0%    25.8MB ± 0%    ~     (p=0.226 n=9+8)
Windower/SUM()_OVER_(ORDER_BY)-12                                               33.8MB ± 0%    33.8MB ± 0%    ~     (p=0.623 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/)-12                93.7MB ± 0%    93.7MB ± 0%    ~     (p=0.684 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/_ORDER_BY)-12        118MB ± 0%     118MB ± 0%    ~     (p=0.247 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/)-12             15.9MB ± 0%    15.9MB ± 0%    ~     (p=0.345 n=10+9)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/_ORDER_BY)-12    22.4MB ± 0%    22.4MB ± 0%    ~     (p=0.436 n=10+10)

name                                                                          old allocs/op  new allocs/op  delta
Windower/SUM()_OVER_()-12                                                         300k ± 0%      300k ± 0%    ~     (all equal)
Windower/SUM()_OVER_(ORDER_BY)-12                                                 600k ± 0%      600k ± 0%    ~     (all equal)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/)-12                 1.20M ± 0%     1.20M ± 0%    ~     (p=0.724 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/_ORDER_BY)-12        1.40M ± 0%     1.40M ± 0%    ~     (p=0.271 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/)-12               302k ± 0%      302k ± 0%    ~     (all equal)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/_ORDER_BY)-12      502k ± 0%      502k ± 0%    ~     (p=0.294 n=10+8)
```