@yuzefovich 2019-07-16
I also noticed that there was a bug in benchmarks of windower - `FilterColIdx` was not properly set. As a result, the code "thought" that there was a filter on zeroth column and comparing it to `DBoolTrue`, and since we have ints in there, the comparison failed as if the row was filtered out shortcircuiting the execution.

Memory account fix doesn't have any significant performance difference, but the true speed went down by about 50%:
```
name                                                                          old time/op    new time/op    delta
Windower/SUM()_OVER_()-24                                                        120ms ± 4%     123ms ± 3%  +1.97%  (p=0.011 n=10+10)
Windower/SUM()_OVER_(ORDER_BY)-24                                                215ms ± 2%     219ms ± 2%  +1.66%  (p=0.004 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/)-24                 328ms ± 3%     337ms ± 1%  +2.81%  (p=0.006 n=10+8)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/_ORDER_BY)-24        350ms ± 3%     354ms ± 3%    ~     (p=0.089 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/)-24              147ms ± 1%     150ms ± 1%  +1.97%  (p=0.000 n=10+9)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/_ORDER_BY)-24     184ms ± 1%     185ms ± 1%  +0.91%  (p=0.009 n=10+10)

name                                                                          old speed      new speed      delta
Windower/SUM()_OVER_()-24                                                     20.0MB/s ± 4%  19.6MB/s ± 2%  -1.95%  (p=0.010 n=10+10)
Windower/SUM()_OVER_(ORDER_BY)-24                                             11.2MB/s ± 2%  11.0MB/s ± 2%  -1.62%  (p=0.004 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/)-24              7.32MB/s ± 3%  7.12MB/s ± 1%  -2.77%  (p=0.005 n=10+8)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/_ORDER_BY)-24     6.87MB/s ± 3%  6.79MB/s ± 3%    ~     (p=0.100 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/)-24           16.3MB/s ± 1%  16.0MB/s ± 1%  -1.94%  (p=0.000 n=10+9)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/_ORDER_BY)-24  13.1MB/s ± 1%  13.0MB/s ± 1%  -0.90%  (p=0.007 n=10+10)

name                                                                          old alloc/op   new alloc/op   delta
Windower/SUM()_OVER_()-24                                                       33.0MB ± 0%    33.0MB ± 0%    ~     (p=0.397 n=8+8)
Windower/SUM()_OVER_(ORDER_BY)-24                                               41.0MB ± 0%    41.0MB ± 0%    ~     (p=0.217 n=9+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/)-24                 101MB ± 0%     101MB ± 0%    ~     (p=0.796 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/_ORDER_BY)-24        125MB ± 0%     125MB ± 0%    ~     (p=0.243 n=10+9)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/)-24             23.1MB ± 0%    23.1MB ± 0%    ~     (p=0.579 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/_ORDER_BY)-24    29.6MB ± 0%    29.6MB ± 0%    ~     (p=0.075 n=10+10)

name                                                                          old allocs/op  new allocs/op  delta
Windower/SUM()_OVER_()-24                                                         600k ± 0%      600k ± 0%    ~     (all equal)
Windower/SUM()_OVER_(ORDER_BY)-24                                                 900k ± 0%      900k ± 0%    ~     (all equal)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/)-24                 1.50M ± 0%     1.50M ± 0%    ~     (p=0.782 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/_ORDER_BY)-24        1.70M ± 0%     1.70M ± 0%    ~     (p=0.251 n=10+9)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/)-24               601k ± 0%      601k ± 0%    ~     (p=1.000 n=10+10)
Windower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/_ORDER_BY)-24      801k ± 0%      801k ± 0%  +0.00%  (p=0.043 n=9+10)
```